Proof of concept caching of the HTTP call
This is a simple way to cache the API request in
a way that would be useable on a multi-user system.
There are two 'bits'. First, we make the request for
the 'top' of the half hour (so hh:00:00 or hh:30:00)
by changing the current time before calling the API,
then we make the call through requests_cache rather
than requests. Because the time (and postcode) is
part of the request this means that old data is
reused, avoiding the HTTP call.

Currently the results are just in file `cats_cache.sqlits`
in the current working directory, but we could put this
in a hidden file in the users homespace, or in a global
temporary directory or whatever.