Make sure our timestamps are timezone aware
This probably needs to be done on an api-interface
by api-interface basis (as how we do it depends on
how we convert whatever comes back into a datetime
object). For the ciuk we have explicit UTC returned
but strptime does not know how to parse this. So
add the timezone info for each datetime object.

Also modify the test to make sure all datetime
objects we get back from the api_query are aware
of their datetime (we don't care what this is,
only that it is defined). We can then convert to
system time at the other end.