Report times in the local system timezone
As far as I can tell, when using the posix timestamp
to at, it works in local system time (which for me is
BST). So make all our start times and reports from the
optimiser timezone aware and in local system time. It
turns out that all we need to do to do this is feed in
a start time that is timezone aware and in the right
timezone. Doing this shifts the time that gets passed
generated for feeding into at by 1 hr, which is what we
want.

Also add a test case to test_windowed_forecast to make
sure that this all works. Note that we have to explicitly
check the timezones as the same time with different timezones
compare equal.