Fix errors and issues picked up by type checking
There was a handful of issues where static type
checking either picked up errors or ran into
code that could not verified by the mypy static
type checking. Specifically:

1. in configure.py: Fix a typo in logging (changing
"eror" to "error" that was not covered in our pytest
test cases. This would have resulted in an error
had we hit this condition.

2. in __init__.py: We now always return an integer
from main (so we sys.exit 0 if everything works).
While None being passed into sys.exit isn't an
error in python, it does break static type checking.

3. in __init__.py: We also avoid reusing the variable
name "output" for the subrocess output (bytes) and
the cats output type that goes into the subprocess.
While this isn't an error in python, it does break
static type checking.