Sort out typing problems with check_clean_arguments
But I cannot help but think that this is a long way
around. Unwrapping the dictonary construtor so that
mypy can see the types (aroud line 32) seems more
complex than it should be, and the cast at the end
seems painful. Can we not just build the TypedDict
as we go?