fix(forecast): fixes off-by-one error and bisect_left()
This fixes issues in the current CATS implementation which led to
https://github.com/GreenScheduler/cats/issues/99

Firstly, bisect_left() which finds the index of the closest data point
past the job end time, was looking at the datetime attribute of the
CarbonIntensityPointEstimate object, which by convention only has the
start time of the window for which carbon intensity is reported from
ciuk_parse_response_data(). This led to an error where if the job end
time fell in the last window, bisect_left() would return None as it did
not find a starting datetime past the job end time. This commit fixes
this by looking at the end of the window.

Secondly, this fixes __len__ on WindowedForecast which had a off by one
error on the number of windows. This was triggerred when the number of
windows was the same size as the data which should be valid, but gave an
empty iterable.

Thirdly, fixed a IndexError in rbound estimation that was caused by this
new method. This happened when the job end time fell on the last data
point. Interpolation between this data point and the next one failed, as
there was no succeeding data point. Also corrected off-by-one shifts in
the list of windows for the trapezoidal rule.

Fixes: #99