Reduce number of windows by one in WindowedForecast
This is a partial revert of c08a256f0bdccfdb58c7308be96c7b896a68242b
that added one to the number of windows to ensure that for low durations
the number of windows is the same as the number of data points. While this
fixed errors for durations close to 48h, it introduced #119 that caused
errors for low durations (5 min). This is caused by the interpolation
trying to read past the last data frame.

Fixes: #119