Fix start and end points in plotting
For plotting we were only including points for the optimal and
now start time from within the dataset of point estimates (i.e.
every 30 mins) even if the start or end time was not on one of
these times. This means the plots missed time at the start or
end of the projected run time for the red and green segments.

This is fixed by adding two more fields to the CarbonIntensityAverageEstimate,
the start_value and end_value which correspond to the forcast
values from the start and end times. These are interpolated when
calculating the integral for the average estimate anyway, we just store the
data now. We then pick up and plot these data points at the start and end
of the red and green segments.

This adds information into the serialised output in JSON format and in
the expected CarbonIntensityAverageEstimates in the test cases. Hence the
test cases are updated.