Generate warning, allow config.yaml
We now log (at warning level) the fact that config.yml (and config.yaml,
which is now accepted) are deprecated in favour of cats_config.yml/yaml.

Add a new test to check things for for config.yaml.

To do this in a sensible way, I removed the early returns from within
config_from_file() which resulted in us skipping some existing
logging messages. We now only return at the end of the function and
logging is performed as appeared to be intended in the code. We could
(I think) simplify the logic further by collapsing all the possible
paths into a single loop so we just accept the first. But that's for
later refactoring.