Fixup docs-build-deploy action on main
Our first github actions job when committing on main is to build the docs and push to a branch for github pages. This is only triggered for changes to the main branch. I recently changed the checkout step used for this job to avoid persisting the credentials to disk on the runner (for security) but this stops the push from happening. This fixup patch reverts that case.